chore: export client & types
